DjangoProject/app/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==3.6.4.0
celery==5.2.7
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
defusedxml==0.7.1
Deprecated==1.2.14
Django==5.0.6
django-allauth==0.63.3
django-debug-toolbar==4.4.2
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
flower==1.1.0
humanize==4.9.0
idna==3.7
kombu==5.3.7
oauthlib==3.2.2
packaging==24.0
pillow==10.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.46
psycopg==3.1.12
psycopg-binary==3.1.12
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2024.1
redis==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.0
tornado==6.4.1
typing_extensions==4.12.0
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
gunicorn==21.2.0